add Source and Target MM in the transformation
These metamodels are needed since the ATL model itself does not contain enough
information to perform a correct translation. For example there is no way to
know if a navigation is an attribute access or a reference traversal, which are
mapped differently into Gremlin.

This commit also add a dependency to ClassDiagram, which cannot be resolved by
external users (refs #26). It has to be removed before releasing ATL2Gremlin